requirements.txt
Django==1.9.5
djangorestframework==3.3.1
djangorestframework-bulk==0.2.1
psycopg2==2.6.1
requests==2.8.1
django-cors-headers==1.1.0
django-rest-swagger=0.3.6
django-filter=0.13.0